pyproject.toml
[project]
name = "secure-messenger"
version = "0.1.0"
description = "End-to-end encrypted messenger with zero-knowledge server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "httpx>=0.28.1",
    "pynacl>=1.5.0",
    "uvicorn>=0.34.3",
    "python-multipart>=0.0.9",
    "pydantic>=2.5.0",
]

[project.scripts]
secure-messenger = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server", "client", "crypto"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
